requirements.txt
absl-py==0.9.0
add-one==1.3.0
aniso8601==8.0.0
appdirs==1.4.4
astunparse==1.6.3
attrs==19.3.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
flask-blueprint==1.3.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
future==0.18.2
gast==0.3.3
genomes==1.1.0
google-auth==1.16.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphsurgeon @ file:///home/app/TensorRT-6.0.1.5/graphsurgeon/graphsurgeon-0.4.1-py2.py3-none-any.whl
grpcio==1.29.0
h5py==2.10.0
idna==2.9
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
more-itertools==8.4.0
networkx==2.4
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.4
Pillow==7.1.2
pluggy==0.13.1
protobuf==3.12.2
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pycuda==2019.1.2
pycwt==0.3.0a22
PyMySQL==0.10.0
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-vlc==3.0.10114
pytools==2020.3.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
scikit-image==0.17.2
scipy==1.4.1
Shapely==1.7.1
six==1.14.0
SQLAlchemy==1.3.18
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorrt @ file:///home/app/TensorRT-6.0.1.5/python/tensorrt-6.0.1.5-cp37-none-linux_x86_64.whl
termcolor==1.1.0
terminaltables==3.1.0
test-cpp==0.1
test-cuda===1.x
tifffile==2020.7.4
torch==1.2.0
torchscope @ git+https://github.com/Tramac/torchscope.git@a6afa094919f70b1db85a4afc443b1a1ef1dd21e
torchvision==0.4.0
tqdm==4.46.0
uff @ file:///home/app/TensorRT-6.0.1.5/uff/uff-0.6.5-py2.py3-none-any.whl
urllib3==1.25.9
vision==1.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
